+++ lib/aixm/feature/obstacle.rb
@@ -12,20 +12,29 @@ module AIXM
     #     name: String or nil
     #     type: TYPES
     #     xy: AIXM.xy
-    #     radius: AIXM.d
     #     z: AIXM.z
+    #     radius: AIXM.d
     #   )
     #   obstacle.lighting = true or false (default for AIXM) or nil (means: unknown, default for OFMX)
     #   obstacle.lighting_remarks = String or nil
     #   obstacle.marking = true or false or nil (means: unknown, default)
     #   obstacle.marking_remarks = String or nil
     #   obstacle.height = AIXM.d or nil
+    #   obstacle.height_accurate = true or false or nil (means: unknown, default)
     #   obstacle.xy_accuracy = AIXM.d or nil
     #   obstacle.z_accuracy = AIXM.d or nil
-    #   obstacle.height_accurate = true or false or nil (means: unknown, default)
     #   obstacle.valid_from = Time or Date or String or nil
     #   obstacle.valid_until = Time or Date or String or nil
     #   obstacle.remarks = String or nil
+    #   obstacle.link_to   # => AIXM.obstacle or nil
+    #   obstacle.link_type   # => LINK_TYPE or nil
+    #
+    # See {AIXM::Feature::ObstacleGroup} for how to define physical links
+    # between two obstacles (e.g. cables between powerline towers).
+    #
+    # Please note: As soon as an obstacle is added to an obstacle group, the
+    # +xy_accuracy+ and +z_accuracy+ of the obstacle group overwrite whatever
+    # is set on the individual obstacles!
     #
     # @see https://github.com/openflightmaps/ofmx/wiki/Obstacle
     class Obstacle < Feature
@@ -40,6 +49,12 @@ module AIXM
         TOWER: :tower,
         WINDTURBINE: :wind_turbine,
         OTHER: :other   # specify in remarks
+      }.freeze
+
+      LINK_TYPES = {
+        CABLE: :cable,
+        SOLID: :solid,
+        OTHER: :other
       }.freeze
 
       # @return [String] full name
@@ -74,15 +89,15 @@ module AIXM
       # @return [AIXM::D, nil] height from ground to top point
       attr_reader :height
 
+      # @return [Boolean, nil] height accuracy
+      #   true => height measured, false => height estimated, nil => unknown
+      attr_reader :height_accurate
+
       # @return [AIXM::D, nil] margin of error for circular base center point
       attr_reader :xy_accuracy
 
       # @return [AIXM::D, nil] margin of error for top point
       attr_reader :z_accuracy
-
-      # @return [Boolean, nil] height accuracy
-      #   true => height measured, false => height estimated, nil => unknown
-      attr_reader :height_accurate
 
       # @return [Time, Date, String, nil] effective after this point in time
       attr_reader :valid_from
@@ -93,9 +108,18 @@ module AIXM
       # @return [String, nil] free text remarks
       attr_reader :remarks
 
-      def initialize(source: nil, name: nil, type:, xy:, radius:, z:)
+      # @return [AIXM::Feature::ObstacleGroup] group this obstacle belongs to
+      attr_reader :obstacle_group
+
+      # @return [Symbol, nil] another obstacle to which a physical link exists
+      attr_reader :linked_to
+
+      # @return [Symbol, nil] type of physical link between this and another obstacle
+      attr_reader :link_type
+
+      def initialize(source: nil, name: nil, type:, xy:, z:, radius:)
         super(source: source)
-        self.name, self.type, self.xy, self.radius, self.z = name, type, xy, radius, z
+        self.name, self.type, self.xy, self.z, self.radius = name, type, xy, z, radius
         @lighting = @marking = @height_accurate = false
       end
 
@@ -118,14 +142,14 @@ module AIXM
         @xy = value
       end
 
-      def radius=(value)
-        fail(ArgumentError, "invalid radius") unless value.is_a?(AIXM::D) && value.dist > 0
-        @radius = value
-      end
-
       def z=(value)
         fail(ArgumentError, "invalid z") unless value.is_a?(AIXM::Z) && value.qnh?
         @z = value
+      end
+
+      def radius=(value)
+        fail(ArgumentError, "invalid radius") unless value.nil? || (value.is_a?(AIXM::D) && value.dist > 0)
+        @radius = value
       end
 
       def lighting=(value)
@@ -151,6 +175,11 @@ module AIXM
         @height = value
       end
 
+      def height_accurate=(value)
+        fail(ArgumentError, "invalid height accurate") unless [true, false, nil].include? value
+        @height_accurate = value
+      end
+
       def xy_accuracy=(value)
         fail(ArgumentError, "invalid xy accuracy") unless value.nil? || value.is_a?(AIXM::D)
         @xy_accuracy = value
@@ -159,11 +188,6 @@ module AIXM
       def z_accuracy=(value)
         fail(ArgumentError, "invalid z accuracy") unless value.nil? || value.is_a?(AIXM::D)
         @z_accuracy = value
-      end
-
-      def height_accurate=(value)
-        fail(ArgumentError, "invalid height accurate") unless [true, false, nil].include? value
-        @height_accurate = value
       end
 
       def valid_from=(value)
@@ -178,39 +202,59 @@ module AIXM
         @remarks = value&.to_s
       end
 
-      # @return [Boolean] estimation whether one obstacle (e.g. single tree) or
-      #    a cluster of obstacles (e.g. a few trees)
-      def clustered?
-        height && radius > height
+      def obstacle_group=(value)
+        fail(ArgumentError, "invalid obstacle group") unless value.is_a?(AIXM::Feature::ObstacleGroup)
+        @obstacle_group = value
       end
+      private :obstacle_group=
+
+      def linked_to=(value)
+        fail(ArgumentError, "invalid linked to") unless value.is_a?(AIXM::Feature::Obstacle)
+        @linked_to = value
+      end
+      private :linked_to=
+
+      def link_type=(value)
+        @link_type = LINK_TYPES.lookup(value&.to_s&.to_sym, nil) || fail(ArgumentError, "invalid link type")
+      end
+      private :link_type=
 
       # @return [Boolean] whether part of an obstacle group
       def grouped?
-        respond_to? :group
+        obstacle_group && obstacle_group.obstacles.count > 1
+      end
+
+      # @return [Boolean] whether obstacle is linked to another one
+      def linked?
+        !!linked_to
       end
 
       # @return [String] UID markup
       def to_uid(as: :ObsUid)
+        self.obstacle_group ||= singleton_obstacle_group
         builder = Builder::XmlMarkup.new(indent: 2)
         builder.tag!(as) do |tag|
+          tag << obstacle_group.to_uid.indent(2) if AIXM.ofmx?
           tag.geoLat((xy.lat(AIXM.schema)))
           tag.geoLong((xy.long(AIXM.schema)))
         end
       end
 
       # @return [String] AIXM or OFMX markup
-      def to_xml
+      def to_xml(delegate: true)
+        self.obstacle_group ||= singleton_obstacle_group
+        return obstacle_group.to_xml if delegate && AIXM.ofmx?
         builder = Builder::XmlMarkup.new(indent: 2)
         builder.comment! "Obstacle: [#{type}] #{xy.to_s} #{name}".strip
-        builder.Obs({ source: ((source || (group.source if grouped?)) if AIXM.ofmx?) }.compact) do |obs|
+        builder.Obs do |obs|
           obs << to_uid.indent(2)
           obs.txtName(name) if name
           if AIXM.ofmx?
             obs.codeType(TYPES.key(type).to_s)
           else
             obs.txtDescrType(TYPES.key(type).to_s)
-            obs.codeGroup(grouped? || clustered? ? 'Y' : 'N')
           end
+          obs.codeGroup(grouped? ? 'Y' : 'N')
           if AIXM.ofmx?
             obs.codeLgt(lighting ? 'Y' : 'N') unless lighting.nil?
             obs.codeMarking(marking ? 'Y' : 'N') unless marking.nil?
@@ -220,27 +264,27 @@ module AIXM
           obs.txtDescrLgt(lighting_remarks) if lighting_remarks
           obs.txtDescrMarking(marking_remarks) if marking_remarks
           obs.codeDatum('WGE')
-          if xy_accuracy
-            obs.valGeoAccuracy(xy_accuracy.dist.trim)
-            obs.uomGeoAccuracy(xy_accuracy.unit.upcase.to_s)
+          if AIXM.aixm? && obstacle_group.xy_accuracy
+            obs.valGeoAccuracy(obstacle_group.xy_accuracy.dist.trim)
+            obs.uomGeoAccuracy(obstacle_group.xy_accuracy.unit.upcase.to_s)
           end
           obs.valElev(z.alt)
-          obs.valElevAccuracy(z_accuracy.to_ft.dist.round) if z_accuracy
+          if AIXM.aixm? && obstacle_group.z_accuracy
+            obs.valElevAccuracy(obstacle_group.z_accuracy.to_ft.dist.round)
+          end
           obs.valHgt(height.to_ft.dist.round) if height
+          obs.uomDistVer('FT')
           if AIXM.ofmx? && !height_accurate.nil?
             obs.codeHgtAccuracy(height_accurate ? 'Y' : 'N')
           end
-          obs.uomDistVer('FT')
           if AIXM.ofmx?
-            obs.valRadius(radius.dist.trim)
-            obs.uomRadius(radius.unit.upcase.to_s)
-            if grouped?
-              obs.codeGroupId(group.id)
-              obs.txtGroupName(group.name) if group.name
-              if linked?
-                obs << linked_to.to_uid(as: :ObsUidLink).indent(2)
-                obs.codeLinkType(AIXM::Feature::ObstacleGroup::LINK_TYPES.key(link_type).to_s)
-              end
+            if radius
+              obs.valRadius(radius.dist.trim)
+              obs.uomRadius(radius.unit.upcase.to_s)
+            end
+            if grouped? && linked?
+              obs << linked_to.to_uid(as: :ObsUidLink).indent(2)
+              obs.codeLinkType(LINK_TYPES.key(link_type).to_s)
             end
             obs.datetimeValidWef(valid_from.xmlschema) if valid_from
             obs.datetimeValidTil(valid_until.xmlschema) if valid_until
@@ -249,38 +293,12 @@ module AIXM
         end
       end
 
-      # Extensions used by ObstacleGroup feature
-      module Grouped
-        # @return [AIXM::Feature::ObstacleGroup] group this obstacle belongs to
-        attr_reader :group
+      private
 
-        # @return [AIXM::Feature::Obstacle] another obstacle this one is linked to
-        attr_reader :linked_to
-
-        # @return [Symbol] type of link between two obstacles
-        attr_reader :link_type
-
-        def group=(value)
-          fail(ArgumentError, "invalid group") unless value.is_a?(AIXM::Feature::ObstacleGroup)
-          @group = value
-        end
-        private :group=
-
-        # @return [Boolean] whether obstacle is linked to another one
-        def linked?
-          !!linked_to
-        end
-
-        def linked_to=(value)
-          fail(ArgumentError, "invalid linked to") unless value.is_a?(AIXM::Feature::Obstacle)
-          @linked_to = value
-        end
-        private :linked_to=
-
-        def link_type=(value)
-          @link_type = AIXM::Feature::ObstacleGroup::LINK_TYPES.lookup(value&.to_s&.to_sym, nil) || fail(ArgumentError, "invalid link type")
-        end
-        private :link_type=
+      # OFMX requires single, ungrouped obstacles to be the only member of a
+      # singleton obstacle group.
+      def singleton_obstacle_group
+        AIXM.obstacle_group.add_obstacle self
       end
 
     end