+++ spec/lib/aixm/document_spec.rb
@@ -1339,122 +1339,175 @@ describe AIXM::Document do
             </Ttt>
             <txtRmk>vortac navaid</txtRmk>
           </Tcn>
+          <!-- Obstacle group: EIFFEL TOWER -->
+          <Ogr>
+            <OgrUid>
+              <txtName>EIFFEL TOWER</txtName>
+              <geoLat>48.85825000N</geoLat>
+              <geoLong>002.29458889E</geoLong>
+            </OgrUid>
+            <codeDatum>WGE</codeDatum>
+            <valGeoAccuracy>2</valGeoAccuracy>
+            <uomGeoAccuracy>M</uomGeoAccuracy>
+            <valElevAccuracy>3</valElevAccuracy>
+            <uomElevAccuracy>FT</uomElevAccuracy>
+          </Ogr>
           <!-- Obstacle: [tower] 48.85825000N 002.29458889E EIFFEL TOWER -->
           <Obs>
             <ObsUid>
+              <OgrUid>
+                <txtName>EIFFEL TOWER</txtName>
+                <geoLat>48.85825000N</geoLat>
+                <geoLong>002.29458889E</geoLong>
+              </OgrUid>
               <geoLat>48.85825000N</geoLat>
               <geoLong>002.29458889E</geoLong>
             </ObsUid>
             <txtName>EIFFEL TOWER</txtName>
             <codeType>TOWER</codeType>
+            <codeGroup>N</codeGroup>
             <codeLgt>Y</codeLgt>
             <txtDescrLgt>red strobes</txtDescrLgt>
             <codeDatum>WGE</codeDatum>
-            <valGeoAccuracy>2</valGeoAccuracy>
-            <uomGeoAccuracy>M</uomGeoAccuracy>
             <valElev>1187</valElev>
-            <valElevAccuracy>3</valElevAccuracy>
             <valHgt>1063</valHgt>
-            <codeHgtAccuracy>Y</codeHgtAccuracy>
             <uomDistVer>FT</uomDistVer>
+            <codeHgtAccuracy>Y</codeHgtAccuracy>
             <valRadius>88</valRadius>
             <uomRadius>M</uomRadius>
             <datetimeValidWef>2018-01-01T12:00:00+01:00</datetimeValidWef>
             <datetimeValidTil>2019-01-01T12:00:00+01:00</datetimeValidTil>
             <txtRmk>Temporary light installations (white strobes, gyro light etc)</txtRmk>
           </Obs>
+          <!-- Obstacle group: MIRMANDE EOLIENNES -->
+          <Ogr>
+            <OgrUid>
+              <txtName>MIRMANDE EOLIENNES</txtName>
+              <geoLat>44.67501389N</geoLat>
+              <geoLong>004.87256667E</geoLong>
+            </OgrUid>
+            <codeDatum>WGE</codeDatum>
+            <valGeoAccuracy>50</valGeoAccuracy>
+            <uomGeoAccuracy>M</uomGeoAccuracy>
+            <valElevAccuracy>33</valElevAccuracy>
+            <uomElevAccuracy>FT</uomElevAccuracy>
+            <txtRmk>Extension planned</txtRmk>
+          </Ogr>
           <!-- Obstacle: [wind_turbine] 44.67501389N 004.87256667E LA TEISSONIERE 1 -->
           <Obs>
             <ObsUid>
+              <OgrUid>
+                <txtName>MIRMANDE EOLIENNES</txtName>
+                <geoLat>44.67501389N</geoLat>
+                <geoLong>004.87256667E</geoLong>
+              </OgrUid>
               <geoLat>44.67501389N</geoLat>
               <geoLong>004.87256667E</geoLong>
             </ObsUid>
             <txtName>LA TEISSONIERE 1</txtName>
             <codeType>WINDTURBINE</codeType>
+            <codeGroup>Y</codeGroup>
             <codeLgt>N</codeLgt>
             <codeMarking>N</codeMarking>
             <codeDatum>WGE</codeDatum>
-            <valGeoAccuracy>50</valGeoAccuracy>
-            <uomGeoAccuracy>M</uomGeoAccuracy>
             <valElev>1764</valElev>
-            <valElevAccuracy>33</valElevAccuracy>
             <valHgt>262</valHgt>
-            <codeHgtAccuracy>N</codeHgtAccuracy>
             <uomDistVer>FT</uomDistVer>
+            <codeHgtAccuracy>N</codeHgtAccuracy>
             <valRadius>80</valRadius>
             <uomRadius>M</uomRadius>
-            <codeGroupId>462c8d00-981a-0995-09aa-4ba39161bb41</codeGroupId>
-            <txtGroupName>MIRMANDE EOLIENNES</txtGroupName>
           </Obs>
           <!-- Obstacle: [wind_turbine] 44.67946667N 004.87381111E LA TEISSONIERE 2 -->
           <Obs>
             <ObsUid>
+              <OgrUid>
+                <txtName>MIRMANDE EOLIENNES</txtName>
+                <geoLat>44.67501389N</geoLat>
+                <geoLong>004.87256667E</geoLong>
+              </OgrUid>
               <geoLat>44.67946667N</geoLat>
               <geoLong>004.87381111E</geoLong>
             </ObsUid>
             <txtName>LA TEISSONIERE 2</txtName>
             <codeType>WINDTURBINE</codeType>
+            <codeGroup>Y</codeGroup>
             <codeLgt>N</codeLgt>
             <codeMarking>N</codeMarking>
             <codeDatum>WGE</codeDatum>
-            <valGeoAccuracy>50</valGeoAccuracy>
-            <uomGeoAccuracy>M</uomGeoAccuracy>
             <valElev>1738</valElev>
-            <valElevAccuracy>33</valElevAccuracy>
             <valHgt>262</valHgt>
-            <codeHgtAccuracy>N</codeHgtAccuracy>
             <uomDistVer>FT</uomDistVer>
+            <codeHgtAccuracy>N</codeHgtAccuracy>
             <valRadius>80</valRadius>
             <uomRadius>M</uomRadius>
-            <codeGroupId>462c8d00-981a-0995-09aa-4ba39161bb41</codeGroupId>
-            <txtGroupName>MIRMANDE EOLIENNES</txtGroupName>
           </Obs>
+          <!-- Obstacle group: DROITWICH LONGWAVE ANTENNA -->
+          <Ogr>
+            <OgrUid>
+              <txtName>DROITWICH LONGWAVE ANTENNA</txtName>
+              <geoLat>52.29639722N</geoLat>
+              <geoLong>002.10675278W</geoLong>
+            </OgrUid>
+            <codeDatum>WGE</codeDatum>
+            <valGeoAccuracy>0</valGeoAccuracy>
+            <uomGeoAccuracy>M</uomGeoAccuracy>
+            <valElevAccuracy>0</valElevAccuracy>
+            <uomElevAccuracy>FT</uomElevAccuracy>
+            <txtRmk>Destruction planned</txtRmk>
+          </Ogr>
           <!-- Obstacle: [mast] 52.29639722N 002.10675278W DROITWICH LW NORTH -->
           <Obs>
             <ObsUid>
+              <OgrUid>
+                <txtName>DROITWICH LONGWAVE ANTENNA</txtName>
+                <geoLat>52.29639722N</geoLat>
+                <geoLong>002.10675278W</geoLong>
+              </OgrUid>
               <geoLat>52.29639722N</geoLat>
               <geoLong>002.10675278W</geoLong>
             </ObsUid>
             <txtName>DROITWICH LW NORTH</txtName>
             <codeType>MAST</codeType>
+            <codeGroup>Y</codeGroup>
             <codeLgt>N</codeLgt>
             <codeMarking>N</codeMarking>
             <codeDatum>WGE</codeDatum>
-            <valGeoAccuracy>0</valGeoAccuracy>
-            <uomGeoAccuracy>M</uomGeoAccuracy>
             <valElev>848</valElev>
-            <valElevAccuracy>0</valElevAccuracy>
             <valHgt>700</valHgt>
-            <codeHgtAccuracy>Y</codeHgtAccuracy>
             <uomDistVer>FT</uomDistVer>
+            <codeHgtAccuracy>Y</codeHgtAccuracy>
             <valRadius>200</valRadius>
             <uomRadius>M</uomRadius>
-            <codeGroupId>18e65683-798d-0941-8de4-cb65a6427035</codeGroupId>
-            <txtGroupName>DROITWICH LONGWAVE ANTENNA</txtGroupName>
           </Obs>
           <!-- Obstacle: [mast] 52.29457778N 002.10568611W DROITWICH LW NORTH -->
           <Obs>
             <ObsUid>
+              <OgrUid>
+                <txtName>DROITWICH LONGWAVE ANTENNA</txtName>
+                <geoLat>52.29639722N</geoLat>
+                <geoLong>002.10675278W</geoLong>
+              </OgrUid>
               <geoLat>52.29457778N</geoLat>
               <geoLong>002.10568611W</geoLong>
             </ObsUid>
             <txtName>DROITWICH LW NORTH</txtName>
             <codeType>MAST</codeType>
+            <codeGroup>Y</codeGroup>
             <codeLgt>N</codeLgt>
             <codeMarking>N</codeMarking>
             <codeDatum>WGE</codeDatum>
-            <valGeoAccuracy>0</valGeoAccuracy>
-            <uomGeoAccuracy>M</uomGeoAccuracy>
             <valElev>848</valElev>
-            <valElevAccuracy>0</valElevAccuracy>
             <valHgt>700</valHgt>
-            <codeHgtAccuracy>Y</codeHgtAccuracy>
             <uomDistVer>FT</uomDistVer>
+            <codeHgtAccuracy>Y</codeHgtAccuracy>
             <valRadius>200</valRadius>
             <uomRadius>M</uomRadius>
-            <codeGroupId>18e65683-798d-0941-8de4-cb65a6427035</codeGroupId>
-            <txtGroupName>DROITWICH LONGWAVE ANTENNA</txtGroupName>
             <ObsUidLink>
+              <OgrUid>
+                <txtName>DROITWICH LONGWAVE ANTENNA</txtName>
+                <geoLat>52.29639722N</geoLat>
+                <geoLong>002.10675278W</geoLong>
+              </OgrUid>
               <geoLat>52.29639722N</geoLat>
               <geoLong>002.10675278W</geoLong>
             </ObsUidLink>