+++ spec/lib/aixm/feature/obstacle_group_spec.rb
@@ -1,7 +1,7 @@
 require_relative '../../../spec_helper'
 
 describe AIXM::Feature::ObstacleGroup do
-  describe "unlinked group" do
+  describe "unlinked obstacle group" do
     subject do
       AIXM.obstacle_group(name: "Mirmande éoliennes")
     end
@@ -25,30 +25,35 @@ describe AIXM::Feature::ObstacleGroup do
       end
     end
 
-    describe :add_obstacle do
-      it "adds an obstacle to the group" do
-        subject.add_obstacle(AIXM::Factory.obstacle)
-        subject.obstacles.count.must_equal 1
-        subject.obstacles.first.must_be :grouped?
-        subject.obstacles.first.group.must_equal subject
+    describe :xy_accuracy= do
+      it "fails on invalid values" do
+        [:foobar, 123].wont_be_written_to subject, :xy_accuracy
+      end
+
+      it "accepts valid values" do
+        [nil, AIXM::Factory.d, AIXM.d(0, :m)].must_be_written_to subject, :xy_accuracy
       end
     end
 
-    describe :id do
-      subject do
-        AIXM::Factory.unlinked_obstacle_group
+    describe :z_accuracy= do
+      it "fails on invalid values" do
+        [:foobar, 123].wont_be_written_to subject, :z_accuracy
       end
 
-      it "is derived from the group name" do
-        subject.id.must_equal '462c8d00-981a-0995-09aa-4ba39161bb41'
-        subject.name = 'Oggy'
-        subject.id.must_equal '939de24d-941b-d5a6-3437-df135ddb9906'
+      it "accepts valid values" do
+        [nil, AIXM::Factory.d, AIXM.d(0, :m)].must_be_written_to subject, :z_accuracy
       end
+    end
 
-      it "is derived from the xy of the group obstacles" do
-        subject.id.must_equal '462c8d00-981a-0995-09aa-4ba39161bb41'
-        subject.obstacles.first.xy.long = 1
-        subject.id.must_equal '819d6977-5301-14d3-6d2e-1de007d922e1'
+    describe :remarks= do
+      macro :remarks
+    end
+
+    describe :add_obstacle do
+      it "adds an obstacle to the obstacle group" do
+        subject.add_obstacle(AIXM::Factory.obstacle)
+        subject.obstacles.count.must_equal 1
+        subject.obstacles.first.obstacle_group.must_equal subject
       end
     end
 
@@ -102,64 +107,80 @@ describe AIXM::Feature::ObstacleGroup do
       it "builds correct OFMX" do
         AIXM.ofmx!
         subject.to_xml.must_equal <<~END
+          <!-- Obstacle group: MIRMANDE EOLIENNES -->
+          <Ogr>
+            <OgrUid>
+              <txtName>MIRMANDE EOLIENNES</txtName>
+              <geoLat>44.67501389N</geoLat>
+              <geoLong>004.87256667E</geoLong>
+            </OgrUid>
+            <codeDatum>WGE</codeDatum>
+            <valGeoAccuracy>50</valGeoAccuracy>
+            <uomGeoAccuracy>M</uomGeoAccuracy>
+            <valElevAccuracy>33</valElevAccuracy>
+            <uomElevAccuracy>FT</uomElevAccuracy>
+            <txtRmk>Extension planned</txtRmk>
+          </Ogr>
           <!-- Obstacle: [wind_turbine] 44.67501389N 004.87256667E LA TEISSONIERE 1 -->
           <Obs>
             <ObsUid>
+              <OgrUid>
+                <txtName>MIRMANDE EOLIENNES</txtName>
+                <geoLat>44.67501389N</geoLat>
+                <geoLong>004.87256667E</geoLong>
+              </OgrUid>
               <geoLat>44.67501389N</geoLat>
               <geoLong>004.87256667E</geoLong>
             </ObsUid>
             <txtName>LA TEISSONIERE 1</txtName>
             <codeType>WINDTURBINE</codeType>
+            <codeGroup>Y</codeGroup>
             <codeLgt>N</codeLgt>
             <codeMarking>N</codeMarking>
             <codeDatum>WGE</codeDatum>
-            <valGeoAccuracy>50</valGeoAccuracy>
-            <uomGeoAccuracy>M</uomGeoAccuracy>
             <valElev>1764</valElev>
-            <valElevAccuracy>33</valElevAccuracy>
             <valHgt>262</valHgt>
-            <codeHgtAccuracy>N</codeHgtAccuracy>
             <uomDistVer>FT</uomDistVer>
+            <codeHgtAccuracy>N</codeHgtAccuracy>
             <valRadius>80</valRadius>
             <uomRadius>M</uomRadius>
-            <codeGroupId>462c8d00-981a-0995-09aa-4ba39161bb41</codeGroupId>
-            <txtGroupName>MIRMANDE EOLIENNES</txtGroupName>
           </Obs>
           <!-- Obstacle: [wind_turbine] 44.67946667N 004.87381111E LA TEISSONIERE 2 -->
           <Obs>
             <ObsUid>
+              <OgrUid>
+                <txtName>MIRMANDE EOLIENNES</txtName>
+                <geoLat>44.67501389N</geoLat>
+                <geoLong>004.87256667E</geoLong>
+              </OgrUid>
               <geoLat>44.67946667N</geoLat>
               <geoLong>004.87381111E</geoLong>
             </ObsUid>
             <txtName>LA TEISSONIERE 2</txtName>
             <codeType>WINDTURBINE</codeType>
+            <codeGroup>Y</codeGroup>
             <codeLgt>N</codeLgt>
             <codeMarking>N</codeMarking>
             <codeDatum>WGE</codeDatum>
-            <valGeoAccuracy>50</valGeoAccuracy>
-            <uomGeoAccuracy>M</uomGeoAccuracy>
             <valElev>1738</valElev>
-            <valElevAccuracy>33</valElevAccuracy>
             <valHgt>262</valHgt>
-            <codeHgtAccuracy>N</codeHgtAccuracy>
             <uomDistVer>FT</uomDistVer>
+            <codeHgtAccuracy>N</codeHgtAccuracy>
             <valRadius>80</valRadius>
             <uomRadius>M</uomRadius>
-            <codeGroupId>462c8d00-981a-0995-09aa-4ba39161bb41</codeGroupId>
-            <txtGroupName>MIRMANDE EOLIENNES</txtGroupName>
           </Obs>
         END
       end
     end
   end
 
-  describe "linked group" do
+  describe "linked obstacle group" do
     subject do
       AIXM.obstacle_group(name: "Mirmande éoliennes")
     end
 
     describe :add_obstacle do
-      it "adds an obstacle to the group and links it to previous" do
+      it "adds an obstacle to the obstacle group and links it to previous" do
         subject.add_obstacle(AIXM::Factory.obstacle)
         subject.add_obstacle(AIXM::Factory.obstacle, linked_to: :previous, link_type: :cable)
         subject.obstacles.count.must_equal 2
@@ -167,7 +188,7 @@ describe AIXM::Feature::ObstacleGroup do
         subject.obstacles.last.link_type.must_equal :cable
       end
 
-      it "adds an obstacle to the group and links it to another obstacle" do
+      it "adds an obstacle to the obstacle group and links it to another obstacle" do
         subject.add_obstacle(AIXM::Factory.obstacle)
         subject.add_obstacle(AIXM::Factory.obstacle, linked_to: subject.obstacles.first, link_type: :solid)
         subject.obstacles.count.must_equal 2
@@ -226,52 +247,73 @@ describe AIXM::Feature::ObstacleGroup do
       it "builds correct OFMX" do
         AIXM.ofmx!
         subject.to_xml.must_equal <<~END
+          <!-- Obstacle group: DROITWICH LONGWAVE ANTENNA -->
+          <Ogr>
+            <OgrUid>
+              <txtName>DROITWICH LONGWAVE ANTENNA</txtName>
+              <geoLat>52.29639722N</geoLat>
+              <geoLong>002.10675278W</geoLong>
+            </OgrUid>
+            <codeDatum>WGE</codeDatum>
+            <valGeoAccuracy>0</valGeoAccuracy>
+            <uomGeoAccuracy>M</uomGeoAccuracy>
+            <valElevAccuracy>0</valElevAccuracy>
+            <uomElevAccuracy>FT</uomElevAccuracy>
+            <txtRmk>Destruction planned</txtRmk>
+          </Ogr>
           <!-- Obstacle: [mast] 52.29639722N 002.10675278W DROITWICH LW NORTH -->
           <Obs>
             <ObsUid>
+              <OgrUid>
+                <txtName>DROITWICH LONGWAVE ANTENNA</txtName>
+                <geoLat>52.29639722N</geoLat>
+                <geoLong>002.10675278W</geoLong>
+              </OgrUid>
               <geoLat>52.29639722N</geoLat>
               <geoLong>002.10675278W</geoLong>
             </ObsUid>
             <txtName>DROITWICH LW NORTH</txtName>
             <codeType>MAST</codeType>
+            <codeGroup>Y</codeGroup>
             <codeLgt>N</codeLgt>
             <codeMarking>N</codeMarking>
             <codeDatum>WGE</codeDatum>
-            <valGeoAccuracy>0</valGeoAccuracy>
-            <uomGeoAccuracy>M</uomGeoAccuracy>
             <valElev>848</valElev>
-            <valElevAccuracy>0</valElevAccuracy>
             <valHgt>700</valHgt>
-            <codeHgtAccuracy>Y</codeHgtAccuracy>
             <uomDistVer>FT</uomDistVer>
+            <codeHgtAccuracy>Y</codeHgtAccuracy>
             <valRadius>200</valRadius>
             <uomRadius>M</uomRadius>
-            <codeGroupId>18e65683-798d-0941-8de4-cb65a6427035</codeGroupId>
-            <txtGroupName>DROITWICH LONGWAVE ANTENNA</txtGroupName>
           </Obs>
           <!-- Obstacle: [mast] 52.29457778N 002.10568611W DROITWICH LW NORTH -->
           <Obs>
             <ObsUid>
+              <OgrUid>
+                <txtName>DROITWICH LONGWAVE ANTENNA</txtName>
+                <geoLat>52.29639722N</geoLat>
+                <geoLong>002.10675278W</geoLong>
+              </OgrUid>
               <geoLat>52.29457778N</geoLat>
               <geoLong>002.10568611W</geoLong>
             </ObsUid>
             <txtName>DROITWICH LW NORTH</txtName>
             <codeType>MAST</codeType>
+            <codeGroup>Y</codeGroup>
             <codeLgt>N</codeLgt>
             <codeMarking>N</codeMarking>
             <codeDatum>WGE</codeDatum>
-            <valGeoAccuracy>0</valGeoAccuracy>
-            <uomGeoAccuracy>M</uomGeoAccuracy>
             <valElev>848</valElev>
-            <valElevAccuracy>0</valElevAccuracy>
             <valHgt>700</valHgt>
-            <codeHgtAccuracy>Y</codeHgtAccuracy>
             <uomDistVer>FT</uomDistVer>
+            <codeHgtAccuracy>Y</codeHgtAccuracy>
             <valRadius>200</valRadius>
             <uomRadius>M</uomRadius>
-            <codeGroupId>18e65683-798d-0941-8de4-cb65a6427035</codeGroupId>
-            <txtGroupName>DROITWICH LONGWAVE ANTENNA</txtGroupName>
             <ObsUidLink>
+              <OgrUid>
+                <txtName>DROITWICH LONGWAVE ANTENNA</txtName>
+                <geoLat>52.29639722N</geoLat>
+                <geoLong>002.10675278W</geoLong>
+              </OgrUid>
               <geoLat>52.29639722N</geoLat>
               <geoLong>002.10675278W</geoLong>
             </ObsUidLink>