+++ spec/lib/aixm/feature/obstacle_spec.rb
@@ -50,11 +50,11 @@ describe AIXM::Feature::Obstacle do
 
   describe :radius= do
     it "fails on invalid values" do
-      [nil, :foobar, 123, AIXM.d(0, :m)].wont_be_written_to subject, :radius
+      [:foobar, 123, AIXM.d(0, :m)].wont_be_written_to subject, :radius
     end
 
     it "accepts valid values" do
-      [AIXM::Factory.d].must_be_written_to subject, :radius
+      [AIXM::Factory.d, nil].must_be_written_to subject, :radius
     end
   end
 
@@ -182,23 +182,24 @@ describe AIXM::Feature::Obstacle do
     macro :remarks
   end
 
-  describe :clustered? do
-    it "returns false if no height is set" do
-      subject.tap { |s| s.height = nil }.wont_be :clustered?
-    end
-
-    it "returns true if radius is bigger than height" do
-      subject.tap { |s| s.radius, s.height = AIXM.d(2, :m), AIXM.d(1, :m) }.must_be :clustered?
-    end
-
-    it "returns false if radius is smaller than height" do
-      subject.tap { |s| s.radius, s.height = AIXM.d(1, :m), AIXM.d(2, :m) }.wont_be :clustered?
+  describe :grouped? do
+    it "returns false since single obstacles are not grouped" do
+      subject.wont_be :grouped?
     end
   end
 
-  describe :grouped? do
-    it "returns false since obstacles are not grouped" do
-      subject.wont_be :grouped?
+  describe :linked? do
+    subject do
+      AIXM::Factory.unlinked_obstacle_group.obstacles.first
+    end
+
+    it "returns false for unlinked obstacles" do
+      subject.wont_be :linked?
+    end
+
+    it "returns true for linked obstacles" do
+      subject = AIXM::Factory.linked_obstacle_group.obstacles.last
+      subject.must_be :linked?
     end
   end
 
@@ -232,24 +233,40 @@ describe AIXM::Feature::Obstacle do
     it "builds correct OFMX" do
       AIXM.ofmx!
       subject.to_xml.must_equal <<~END
+        <!-- Obstacle group: EIFFEL TOWER -->
+        <Ogr>
+          <OgrUid>
+            <txtName>EIFFEL TOWER</txtName>
+            <geoLat>48.85825000N</geoLat>
+            <geoLong>002.29458889E</geoLong>
+          </OgrUid>
+          <codeDatum>WGE</codeDatum>
+          <valGeoAccuracy>2</valGeoAccuracy>
+          <uomGeoAccuracy>M</uomGeoAccuracy>
+          <valElevAccuracy>3</valElevAccuracy>
+          <uomElevAccuracy>FT</uomElevAccuracy>
+        </Ogr>
         <!-- Obstacle: [tower] 48.85825000N 002.29458889E EIFFEL TOWER -->
         <Obs>
           <ObsUid>
+            <OgrUid>
+              <txtName>EIFFEL TOWER</txtName>
+              <geoLat>48.85825000N</geoLat>
+              <geoLong>002.29458889E</geoLong>
+            </OgrUid>
             <geoLat>48.85825000N</geoLat>
             <geoLong>002.29458889E</geoLong>
           </ObsUid>
           <txtName>EIFFEL TOWER</txtName>
           <codeType>TOWER</codeType>
+          <codeGroup>N</codeGroup>
           <codeLgt>Y</codeLgt>
           <txtDescrLgt>red strobes</txtDescrLgt>
           <codeDatum>WGE</codeDatum>
-          <valGeoAccuracy>2</valGeoAccuracy>
-          <uomGeoAccuracy>M</uomGeoAccuracy>
           <valElev>1187</valElev>
-          <valElevAccuracy>3</valElevAccuracy>
           <valHgt>1063</valHgt>
-          <codeHgtAccuracy>Y</codeHgtAccuracy>
           <uomDistVer>FT</uomDistVer>
+          <codeHgtAccuracy>Y</codeHgtAccuracy>
           <valRadius>88</valRadius>
           <uomRadius>M</uomRadius>
           <datetimeValidWef>2018-01-01T12:00:00+01:00</datetimeValidWef>
@@ -257,29 +274,6 @@ describe AIXM::Feature::Obstacle do
           <txtRmk>Temporary light installations (white strobes, gyro light etc)</txtRmk>
         </Obs>
       END
-    end
-  end
-end
-
-describe AIXM::Feature::Obstacle::Grouped do
-  subject do
-    AIXM::Factory.unlinked_obstacle_group.obstacles.first
-  end
-
-  describe :grouped? do
-    it "returns true since obstacles are grouped" do
-      subject.must_be :grouped?
-    end
-  end
-
-  describe :linked? do
-    it "returns false for unlinked obstacles" do
-      subject.wont_be :linked?
-    end
-
-    it "returns true for linked obstacles" do
-      subject = AIXM::Factory.linked_obstacle_group.obstacles.last
-      subject.must_be :linked?
     end
   end
 end